rename files to index, so it has beter naming confensions;
