+++ src/scripts/components/inputText/index.js
