+++ src/scripts/components/loader/index.js
