+++ src/scripts/components/textareaAutosize/index.js
